выполняет 5 задание
